Move the connect form where more than settings can reach it
